+++ x/foundation/msgs_test.go
@@ -755,10 +755,10 @@ func TestMsgSubmitProposalAminoJSON(t *testing.T) {
 		},
 		"MsgUpdateCensorship": {
 			&foundation.MsgUpdateCensorship{
-				Authority:  addrs[0].String(),
+				Authority: addrs[0].String(),
 				Censorship: foundation.Censorship{
 					MsgTypeUrl: sdk.MsgTypeURL((*foundation.MsgWithdrawFromTreasury)(nil)),
-					Authority: foundation.CensorshipAuthorityGovernance,
+					Authority:  foundation.CensorshipAuthorityGovernance,
 				},
 			},
 			fmt.Sprintf("{\"account_number\":\"1\",\"chain_id\":\"foo\",\"fee\":{\"amount\":[],\"gas\":\"0\"},\"memo\":\"memo\",\"msgs\":[{\"type\":\"lbm-sdk/MsgSubmitProposal\",\"value\":{\"exec\":1,\"messages\":[{\"type\":\"lbm-sdk/MsgUpdateCensorship\",\"value\":{\"authority\":\"%s\",\"censorship\":{\"authority\":1,\"msg_type_url\":\"/lbm.foundation.v1.MsgWithdrawFromTreasury\"}}}],\"metadata\":\"MsgUpdateCensorship\",\"proposers\":[\"%s\"]}}],\"sequence\":\"1\",\"timeout_height\":\"1\"}", addrs[0].String(), proposer.String()),